chore: refactor search-index-consumer to improve redability
